refactor: Falldown Shooter を features/ 配下に分割
- FallingShooterPage.tsx（1,602行）を20ファイルに分割
- 型定義、定数、ユーティリティ、Audio、ゲームロジック各モジュールを分離
- UIコンポーネント群（CellView, BulletView, PlayerShip, SkillGauge等）を分離
- カスタムフック（useInterval, useKeyboard, useIdleTimer）を分離
- ユニットテスト7スイート（58テスト）を新規追加
- FallingShooterPage.tsxを薄いラッパー（11行）に変更
- 全100スイート/1,202テスト通過、ビルド成功を確認
