+++ src/features/falldown-shooter/index.ts
@@ -1,0 +1,37 @@
+// 落ち物シューティング barrel export
+
+export { FalldownShooterGame } from './FalldownShooterGame';
+
+// ロジックモジュール
+export { Grid } from './grid';
+export { Block } from './block';
+export { Bullet } from './bullet';
+export { Collision } from './collision';
+export { GameLogic } from './game-logic';
+export { Stage } from './stage';
+export { Audio } from './audio';
+
+// 定数・型・ユーティリティ
+export { CONFIG, BLOCK_COLORS, BLOCK_SHAPES, POWER_TYPES, SKILLS, DEMO_SLIDES } from './constants';
+export { uid, pick, calcTiming } from './utils';
+export type {
+  PowerType,
+  SkillType,
+  GameStatus,
+  OscillatorType,
+  TimingConfig,
+  Config,
+  PowerTypeInfo,
+  SkillInfo,
+  Cell,
+  BlockData,
+  BulletData,
+  GameState,
+  ExplosionData,
+  ParticleData,
+  Powers,
+  CollisionTarget,
+  BulletProcessResult,
+  DemoSlide,
+  KeyboardHandlers,
+} from './types';
